requirements.txt
Appium-Python-Client
selenium
pytest
allure-pytest
pytest-html
pytest-parallel
pytest-xdist
pytest-rerunfailures
pytest-sugar
pytest-json